--HG--
branch : v0.1.0

+++ db/migrate/20120507302333_create_settings_table.rb
@@ -1,17 +1,24 @@
-  class CreateSettingsTable < ActiveRecord::Migration
+class CreateSettingsTable < ActiveRecord::Migration
   def self.up
-    create_table :settings, :force => true do |t|
-      t.string  :var,         :null => false
-      t.text    :value
-      t.integer :target_id
-      t.string  :target_type, :limit => 30
-      t.timestamps
+    if not ActiveRecord::Base.connection.table_exists? 'settings' then
+      create_table :settings, :force => true do |t|
+        t.string  :var,         :null => false
+        t.text    :value
+        t.integer :target_id
+        t.string  :target_type, :limit => 30
+        t.timestamps
+      end
+   
+      add_index :settings, [ :target_type, :target_id, :var ], :unique => true
+
     end
 
-    add_index :settings, [ :target_type, :target_id, :var ], :unique => true
   end
 
   def self.down
-    drop_table :settings
+    if ActiveRecord::Base.connection.table_exists? 'settings' then
+
+      drop_table :settings
+    end
   end
 end
