+++ db/migrate/20150209032053_add_taggings_counter_cache_to_tags.acts_as_taggable_on_engine.rb
@@ -1,8 +1,10 @@
 # This migration comes from acts_as_taggable_on_engine (originally 3)
 class AddTaggingsCounterCacheToTags < ActiveRecord::Migration
   def self.up
-    add_column :tags, :taggings_count, :integer, default: 0
-
+    if !ActiveRecord::Base.connection.column_exists?(:tags, :taggings_count) then
+      add_column :tags, :taggings_count, :integer, default: 0
+    end
+   
     ActsAsTaggableOn::Tag.reset_column_information
     ActsAsTaggableOn::Tag.find_each do |tag|
       ActsAsTaggableOn::Tag.reset_counters(tag.id, :taggings)
@@ -10,6 +12,11 @@ class AddTaggingsCounterCacheToTags < ActiveRecord::Migration
   end
 
   def self.down
-    remove_column :tags, :taggings_count
+    if ActiveRecord::Base.connection.table_exists? :tags then
+
+      if ActiveRecord::Base.connection.column_exists?(:tags, :taggings_count)
+        remove_column :tags, :taggings_count
+      end
+    end
   end
 end
